pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
asyncio_default_fixture_loop_scope = "function"

[tool.poetry]
name = "fastapi-project-template"
version = "0.1.0"
description = "FastAPI Project Template"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
pydantic = "^1.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.18.0"
pytest-cov = "^4.1.0"
black = "^21.7b0"
isort = "^5.9.3"
mypy = "^0.910"
testcontainers = {extras = ["postgresql"], version = "^3.7.0"}
docker = "^6.0.0"
httpx = "^0.24.0" 